[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14160 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringBuilder output = new StringBuilder();
         StringTokenizer token = new StringTokenizer(br.readLine(), " ");
-        int A = Integer.parseInt(token.nextToken());
-        int B = Integer.parseInt(token.nextToken());
+        long A = Integer.parseInt(token.nextToken());
+        long B = Integer.parseInt(token.nextToken());
 
         output.append((A + B) * (A - B));
